add PortfolioWalletHeader, refactor PageHeader to Header
