+++ src/2-pages/portfolio/ui/PortfolioWalletListPage.tsx
@@ -1,11 +1,12 @@
-import {PageHeader, PageWidgetsWrapper} from '@pages/ui';
+import {PageWidgetsWrapper} from '@pages/ui';
 import {PortfolioWalletList} from '@widgets/portfolio';
 import {APP_PATH, APP_TEXT} from '@shared/constants';
+import {Header} from '@shared/ui';
 
 export function PortfolioWalletListPage() {
 	return (
 		<>
-			<PageHeader title={APP_TEXT.connectedWallets} backPath={APP_PATH.portfolio.list} />
+			<Header title={APP_TEXT.connectedWallets} backPath={APP_PATH.portfolio.list} />
 			<PageWidgetsWrapper>
 				<PortfolioWalletList />
 			</PageWidgetsWrapper>
