+++ src/3-widgets/portfolio/index.ts
@@ -1,5 +1,7 @@
 export * from './portfolio-management/ui/PortfolioManagement.tsx';
 export * from './portfolio-wallets/ui/PortfolioWallets.tsx';
 export * from './portfolio-wallet-list/ui/PortfolioWalletList.tsx';
+export * from './portfolio-wallet-details/ui/PortfolioWalletDetails.tsx';
+export * from './portfolio-wallet-header/ui/PortfolioWalletHeader.tsx';
 
 export * from './util/portfolio.constants.ts';
