+++ src/5-shared/ui/index.ts
@@ -28,6 +28,7 @@ export * from './card/ui/Card.tsx';
 export * from './management/ui/Management.tsx';
 export * from './management/type/Management.types.ts';
 export * from './details/ui/Details.tsx';
+export * from './header/ui/Header.tsx';
 
 // new pattern -> get rid of index.ts
